Server/requirements.txt
flask
flask-cors
keras.models
tensorflow
pandas
scikit-learn
numpy
Response
